# swagger/oas30_gen.json
{"openapi":"3.0.0","info":{"title":"srv-episodes","version":"1.0.0","description":"Episodes Service for AnyPodcast"},"paths":{"/episode":{"get":{"operationId":"get_all","responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EpisodeList"}}}}},"parameters":[{"name":"PodcastID","in":"query","required":true,"schema":{"type":"string"}},{"name":"SearchValue","in":"query","required":false,"schema":{"type":"string"}}],"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"post":{"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Episode"}}},"required":true},"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}}},"/episode/{id}":{"get":{"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Episode"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"put":{"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Episode"}}},"required":true},"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"delete":{"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}}}},"components":{"schemas":{"Empty":{"title":"Empty Schema","type":"object"},"Episode":{"type":"object","properties":{"ID":{"type":"number","default":0},"Created":{"type":"string","default":""},"PodcastID":{"type":"string","default":""},"Title":{"type":"string","default":""},"Description":{"type":"string","default":""},"IsVisible":{"type":"boolean","default":false},"Media":{"$ref":"#/components/schemas/Media"}},"required":["Created","Description","ID","IsVisible","Media","PodcastID","Title"],"title":"Episode"},"EpisodeList":{"type":"array","items":{"$ref":"#/components/schemas/Episode"},"title":"EpisodeList"},"Media":{"type":"object","properties":{"CustomName":{"type":"string","default":""},"MediaID":{"type":"string","default":""}},"required":["CustomName","MediaID"],"title":"Media"},"SearchParams":{"type":"object","properties":{"PodcastID":{"type":"string"},"SearchValue":{"type":"string"}},"required":["PodcastID"],"title":"SearchParams"}}},"tags":[],"servers":[],"x-amazon-apigateway-request-validators":{"Validate query string parameters and headers":{"validateRequestParameters":true,"validateRequestBody":false}}}